+++ packages/rand/src/rand.js
@@ -5,8 +5,8 @@ export const rand = l => ~~(random() * l)
 /**
  * From [min, max) return a random integer.
  * Of [min, max), min is inclusive but max is exclusive.
- * @param {number} lo(inclusive) - int
- * @param {number} hi(exclusive) - int
+ * @param {number} lo (inclusive) - int
+ * @param {number} hi (exclusive) - int
  * @returns {number} int
  */
 export const randIn = (lo, hi) => rand(hi - lo) + lo
@@ -14,11 +14,11 @@ export const randIn = (lo, hi) => rand(hi - lo) + lo
 /**
  * From [min, max] return a random integer.
  * Of [min, max], both min and max are inclusive.
- * @param {number} lo(inclusive) - int
- * @param {number} hi(inclusive) - int
+ * @param {number} lo (inclusive) - int
+ * @param {number} hi (inclusive) - int
  * @returns {number} int
  */
-export const randBetw = (lo, hi) => rand(++hi - lo) + lo
+export const betw = (lo, hi) => rand(++hi - lo) + lo
 
 export const randLong = (digit) => {
   let t = ''
@@ -29,8 +29,8 @@ export const randLong = (digit) => {
 /**
  * From [min, max) return a random integer.
  * Of [min, max), min is inclusive but max is exclusive.
- * @param {number} lo(inclusive) - int
- * @param {number} hi(exclusive) - int
+ * @param {number} lo (inclusive) - int
+ * @param {number} hi (exclusive) - int
  * @deprecated use randIn instead
  * @returns {number} int
  */
@@ -39,8 +39,8 @@ export const randInt = (lo, hi) => rand(hi - lo) + lo
 /**
  * From [min, max] return a random integer.
  * Of [min, max], both min and max are inclusive.
- * @param {number} lo(inclusive) - int
- * @param {number} hi(inclusive) - int
+ * @param {number} lo (inclusive) - int
+ * @param {number} hi (inclusive) - int
  * @deprecated use randBetw instead
  * @returns {number} int
  */