+++ packages/rand/test/flopGenerator.test.js
@@ -1,7 +1,7 @@
 import { Foba }          from '@foba/vector-number'
 import { deco }          from '@spare/deco'
 import { logger }        from '@spare/logger'
-import { flopGenerator } from '../src/flopGenerator'
+import { flopGenerator } from '../src/flopGenerator.js'
 
 const arr = Foba.range(5)
 const fg = flopGenerator(arr, -1)