+++ packages/rand/test/ziggurat.test.js
@@ -1,9 +1,16 @@
-import { Histo }              from '@aryth/histo'
-import { decoObject, logger } from '@spare/logger'
-import { Ziggurat }           from '../src/ziggurat'
+import { Histo }                from '@aryth/histo'
+import { decoObject, logger }   from '@spare/logger'
+import { test }                 from 'node:test'
+import { Norm }                 from '../src/normDist/Norm.js'
+import { distributionDemoTape } from './renderer/distributionDemoTape.js'
 
-const zig = Ziggurat.build(0, 10)
-const vec = []
-for (let i = 0; i < 200; i++) vec.push(zig.nextInt())
-const histo = Histo.fromSamples(vec)
-histo.statistics() |> decoObject |> logger
+const gen = Norm.build()
+// const vec = []
+// for (let i = 0; i < 300; i++) vec.push(zig.next())
+const numbers = Array.from({ length: 100000 }, gen.next.bind(gen))
+const histo = Histo.fromSamples(numbers)
+
+test('ziggurat', () => {
+  logger(decoObject(histo.statistics()))
+  distributionDemoTape(numbers, 0.2)
+})