+++ packages/rand/test/zigguratGenerator.test.js
@@ -1,15 +1,17 @@
 import { Histo }              from '@aryth/histo'
-import { deco }               from '@spare/deco'
 import { decoObject, logger } from '@spare/logger'
-import { ziggurat }           from '../src/zigguratGenerator'
+import { test } from 'node:test'
+import { norm } from '../src/normDist/generator.js'
 
-const zig = ziggurat(0, 10, 0)
+const zig = norm(0, 10, 0)
 const vec = []
-for (let i = 0; i < 10; i++) {
+for (let i = 0; i < 300; i++) {
   const norm = zig.next()
-  norm |> deco |> logger
+  // logger(deco(norm))
   vec.push(norm.value)
 }
 
-const histo = Histo.fromSamples(vec)
-histo.statistics() |> decoObject |> logger
+test('ziggurat', () => {
+  const histo = Histo.fromSamples(vec)
+  logger(decoObject(histo.statistics()))
+})